mejorar filtros combinado, agregar estado para limpiar los filtros
