+++ client/src/Components/FilterComponents/brandFilterComponent.jsx
@@ -19,17 +19,14 @@ const BrandFilterComponent = ({ setActualPage }) => {
   };
 
   const handleFilterClick = () => {
-    console.log("Clic en el botón de filtrado");
     try {
       if (!selectedBrand) {
         return;
       }
 
-      console.log("Filtrando por marca:", selectedBrand);
       setActualPage(1);
-      console.log("Filtrando por marca:", selectedBrand);
+
       dispatch(filterBrand(selectedBrand));
-      console.log("Estado después del filtrado:", selectedBrand);
     } catch (error) {
       console.error("Error filtering by brand:", error.message);
     }
@@ -37,36 +34,38 @@ const BrandFilterComponent = ({ setActualPage }) => {
 
   return (
     <div className="mb-4 w-full">
-    <div style={{ marginRight: '4px' }}>
-      <h2 className="block text-sm font-bold mb-4">Brand Filter</h2>
-      <div className="mb-4">
-        <label htmlFor="brand" className="block text-sm font-medium text-gray-700">
-          Seleccionar Marca:
-        </label>
-        <select
-          id="brand"
-          value={selectedBrand}
-          onChange={(e) => handleBrandChange(e.target.value)}
-          className="mt-1 block w-full py-2 px-3 border border-gray-300 bg-white rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
+      <div style={{ marginRight: "4px" }}>
+        <h2 className="block text-sm font-bold mb-4">Brand Filter</h2>
+        <div className="mb-4">
+          <label
+            htmlFor="brand"
+            className="block text-sm font-medium text-gray-700"
+          >
+            Seleccionar Marca:
+          </label>
+          <select
+            id="brand"
+            value={selectedBrand}
+            onChange={(e) => handleBrandChange(e.target.value)}
+            className="mt-1 block w-full py-2 px-3 border border-gray-300 bg-white rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
+          >
+            <option value="">Seleccionar Marca</option>
+            {brandList.map((brand) => (
+              <option key={brand.id} value={brand.brand}>
+                {brand.brand}
+              </option>
+            ))}
+          </select>
+        </div>
+        <button
+          onClick={handleFilterClick}
+          className="transition duration-300 bg-chiliRed hover:bg-onyx text-whiteSmoke font-bold py-2 px-4 rounded"
         >
-          <option value="">Seleccionar Marca</option>
-          {brandList.map((brand) => (
-            <option key={brand.id} value={brand.brand}>
-              {brand.brand}
-            </option>
-          ))}
-        </select>
+          Filtrar
+        </button>
       </div>
-      <button
-        onClick={handleFilterClick}
-        className="transition duration-300 bg-chiliRed hover:bg-onyx text-whiteSmoke font-bold py-2 px-4 rounded"
-      >
-        Filtrar
-      </button>
     </div>
-  </div>
-);
+  );
 };
-
 
 export default BrandFilterComponent;
