+++ client/src/views/Home/Home.jsx
@@ -5,7 +5,7 @@ import FilterPriceComponent from "../../Components/FilterComponents/priceFilterC
 import ColorFilterComponent from "../../Components/FilterComponents/colorFilterComponent";
 import BrandFilterComponent from "../../Components/FilterComponents/brandFilterComponent";
 import SortByPriceButtons from "../../Components/FilterComponents/sortByPriceComponent";
-import AllFilter from "../../Components/FilterComponents/AllFilter";
+
 import Pagination from "../../Components/Pagination/Pagination";
 import { useSelector } from "react-redux";
 
@@ -28,10 +28,16 @@ function Home({ actualPage, handlePageChange, setActualPage }) {
           <div className="mb-2 flex flex-row-reverse">
             <Banners className="mb-2 flex flex-row-reverse" />
             <div className="w-1/5 ml-4 flex flex-col items-start">
-              <AllFilter />
-              {/* <BrandFilterComponent setActualPage={(num) => setActualPage(num)} />
-              <ColorFilterComponent setActualPage={(num) => setActualPage(num)} />
-              <FilterPriceComponent setActualPage={(num) => setActualPage(num)} /> */}
+              {/* <FilterAll setActualPage={(num) => setActualPage(num)} /> */}
+              <BrandFilterComponent
+                setActualPage={(num) => setActualPage(num)}
+              />
+              <ColorFilterComponent
+                setActualPage={(num) => setActualPage(num)}
+              />
+              <FilterPriceComponent
+                setActualPage={(num) => setActualPage(num)}
+              />
               <SortByPriceButtons setActualPage={(num) => setActualPage(num)} />
             </div>
           </div>
